Modify move getToken function to loginActivity
