+++ koin/src/main/java/in/koreatech/koin/ui/login/LoginActivity.kt
@@ -9,6 +9,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
+import com.google.firebase.messaging.FirebaseMessaging
 import dagger.hilt.android.AndroidEntryPoint
 import `in`.koreatech.koin.R
 import `in`.koreatech.koin.common.UiStatus
@@ -51,7 +52,10 @@ class LoginActivity : ActivityBase(R.layout.activity_login) {
                     when (it.status) {
                         is UiStatus.Init -> Unit
                         is UiStatus.Loading -> Unit
-                        is UiStatus.Success -> startActivity(Intent(this@LoginActivity, MainActivity::class.java))
+                        is UiStatus.Success -> {
+                            getToken()
+                            startActivity(Intent(this@LoginActivity, MainActivity::class.java))
+                        }
                         is UiStatus.Failed -> {
                             SnackbarUtil.makeShortSnackbar(binding.root, it.status.message)
                             loginViewModel.onFailedLogin()
@@ -59,6 +63,14 @@ class LoginActivity : ActivityBase(R.layout.activity_login) {
                     }
                 }
             }
+        }
+    }
+
+    private fun getToken() {
+        FirebaseMessaging.getInstance().token.addOnCompleteListener { task ->
+            if (!task.isSuccessful) {}
+            val token = task.result
+            Log.e("device_token", "token : $token")
         }
     }
